Fixed ExtFEMIO for VS.
git-svn-id: https://svnsrv.germanlloyd.org/svn/ExtFEMIO/trunk@331 389f8db8-dd28-11e4-b689-8b400cb78fb4

+++ include/fem/cards_superele.h
@@ -354,7 +354,7 @@ superelement transformation reference number, `ITREF`.
 
                static const dnvgl::extfem::fem::types::entry_type<long> _form_NFIELD;
                static const dnvgl::extfem::fem::types::entry_type<long> _form_ITREF;
-               static const dnvgl::extfem::fem::types::entry_type<long> _form_T;
+               static const dnvgl::extfem::fem::types::entry_type<double> _form_T;
 
             public:
